requirements.txt
Django==1.10
django-extensions==1.6.7
djangorestframework==3.6.2
pkg-resources==0.0.0

six==1.10.0
